requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pymongo==4.6.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.7.8
flask==3.0.0
requests==2.31.0
cryptography==41.0.8
email-validator==2.1.0 